requirements.txt
requests
cloudinary
Pillow
cloudinary